Fusion virtual_*.hpp and virtual_*_*.hpp
